refactor: move and rename getDesktopAppPath
