+++ src/dolphin/ipc.ts
@@ -33,6 +33,12 @@ export const ipc_viewSlpReplay = makeEndpoint.main("viewSlpReplay", <{ files: Re
 
 export const ipc_launchNetplayDolphin = makeEndpoint.main("launchNetplayDolphin", <EmptyPayload>_, <SuccessPayload>_);
 
+export const ipc_getDesktopAppDolphinInfo = makeEndpoint.main(
+  "getDesktopAppDolphinPath",
+  <EmptyPayload>_,
+  <{ dolphinPath: string; exists: boolean }>_,
+);
+
 // toImportDolphin path must point to a "Slippi Dolphin.{exe,app}"
 export const ipc_importDolphinSettings = makeEndpoint.main(
   "importDolphinSettings",
